requirements.txt
requests
pycryptodome
scikit-image
sqlalchemy
shapely
numpy
psutil
dnspython
pydantic
kivy
kivymd
upnpy
tinydb
